requirements.txt
appdirs==1.4.3
astroid==2.3.3
attrs==19.3.0
black==19.10b0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
colorama==0.4.3
Flask==1.1.2
flask-marshmallow==0.11.0
Flask-SocketIO==4.3.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
marshmallow==3.5.1
marshmallow-sqlalchemy==0.22.3
mccabe==0.6.1
mysqlclient==1.4.6
numpy==1.18.2
passlib==1.7.2
pathspec==0.8.0
Pillow==7.1.1
pylint==2.4.4
python-engineio==3.12.1
python-socketio==4.5.1
regex==2020.4.4
requests==2.23.0
sense-hat==2.2.0
six==1.14.0
SQLAlchemy==1.3.16
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.9
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.11.2
WTForms==2.2.1